+++ src/test/java/com/gepardec/recipes/jakarta/JavaxOrmXmlToJakartaOrmXmlTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.gepardec.wor;
+package com.gepardec.recipes.jakarta;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
