+++ src/test/java/com/gepardec/recipes/java/AddAnnotationWithoutArgumentsToClassTest.java
@@ -1,11 +1,6 @@
-package com.gepardec.wor;
+package com.gepardec.recipes.java;
 
-import com.gepardec.wor.lord.wsannotation.AddAnnotationWithoutArgumentsToClass;
 import org.junit.jupiter.api.Test;
-import org.openrewrite.InMemoryExecutionContext;
-import org.openrewrite.ListRuntimeClasspath;
-import org.openrewrite.java.JavaParser;
-import org.openrewrite.test.RecipeSpec;
 import org.openrewrite.test.RewriteTest;
 
 import static org.openrewrite.java.Assertions.java;
@@ -14,7 +9,7 @@ public class AddAnnotationWithoutArgumentsToClassTest implements RewriteTest {
     @Test
     void test() {
         rewriteRun(
-          recipeSpec -> recipeSpec.recipe(new AddAnnotationWithoutArgumentsToClass("Class", "com.gepardec", "org.junit.jupiter.api.Disabled", false)),
+          recipeSpec -> recipeSpec.recipe(new AddAnnotationWithoutArgumentsToClass("com.gepardec.Class", "org.junit.jupiter.api.Disabled", false)),
           java(
             """
               package com.gepardec;
